Move check-font-loaded to actions
