+++ src/actions/check-font-loaded.js
